+++ sphere.cpp
@@ -12,17 +12,16 @@ private:
 
 public:
     float distance_to_surface(float x2, float y2, float z2) override {
-
         float distance = distance_between(x, y, z, x2, y2, z2) - radius;
 
         return distance;
     }
 
     static inline float distance_between(float x1, float y1, float z1, float x2, float y2, float z2) {
-        return sqrt(
-                pow(x1 - x2, 2)
-                + pow(y1 - y2, 2)
-                + pow(z1 - z2, 2));
+        return sqrtf(
+                powf(x1 - x2, 2)
+                + powf(y1 - y2, 2)
+                + powf(z1 - z2, 2));
     }
 
     uint8_t get_color_r() override {
